requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg[binary]==3.2.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
resend==1.4.0
Flask-Bcrypt==1.0.1
email-validator==2.3.0
waitress==3.0.2
whitenoise==6.6.0
gunicorn==21.2.0